+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A new class `WalletFactory` encapsulates functionality for creating a `Wallet` object from a public and private key.
+
 ## [5.0.4] - 2020-05-15
 
 This release contains minor deprecations of names of methods and classes to make this library more idiomatic with the wider JavaScript ecosystem.